Anna: Worked on object pools and basic game functionality
Worked on object pools without a lot of success. When object pools are setup correctly, the IsAlive variable will work correctly, and the tower will not be blocked anymore on the first target.
Added the currency and castle health points. They are not displayed for now but enemies slayn can give currency and enemies reaching the castle can reduce health.